+++ src/test/java/org/atatec/trugger/test/ElementFinderMock.java
@@ -18,15 +18,13 @@ package org.atatec.trugger.test;
 
 import org.atatec.trugger.Finder;
 import org.atatec.trugger.element.Element;
-import org.atatec.trugger.util.mock.MockBuilder;
-import org.easymock.EasyMock;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 
-import static org.easymock.EasyMock.expect;
-import static org.easymock.EasyMock.replay;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 /**
  * A builder for creating {@link Finder} mock objects for {@link Element
@@ -36,16 +34,16 @@ import static org.easymock.EasyMock.replay;
  *
  * @author Marcelo Varella Barca Guimarães
  */
-public class ElementFinderMock implements MockBuilder<Finder<Element>> {
+public class ElementFinderMock {
 
   private final Finder<Element> finder;
   private final Map<String, Element> elements;
 
   public ElementFinderMock() {
-    finder = EasyMock.createMock(Finder.class);
+    finder = mock(Finder.class);
     elements = new HashMap<>();
-    expect(finder.findAll())
-        .andReturn(target -> new ArrayList<>(elements.values())).anyTimes();
+    when(finder.findAll())
+        .thenReturn(target -> new ArrayList<>(elements.values()));
   }
 
   /**
@@ -56,8 +54,8 @@ public class ElementFinderMock implements MockBuilder<Finder<Element>> {
   public ElementFinderMock add(Element... elements) {
     for (Element element : elements) {
       this.elements.put(element.name(), element);
-      expect(finder.find(element.name()))
-          .andReturn(o -> element).anyTimes();
+      when(finder.find(element.name()))
+          .thenReturn(o -> element);
     }
     return this;
   }
@@ -67,14 +65,12 @@ public class ElementFinderMock implements MockBuilder<Finder<Element>> {
    *
    * @return a reference to this object.
    */
-  public ElementFinderMock add(MockBuilder<Element> builder) {
+  public ElementFinderMock add(ElementMock builder) {
     add(builder.createMock());
     return this;
   }
 
-  @Override
   public Finder<Element> createMock() {
-    replay(finder);
     return finder;
   }