+++ src/test/java/org/atatec/trugger/test/reflection/ConstructorsSelectorTest.java
@@ -21,7 +21,7 @@ import org.junit.Test;
 
 import java.util.function.Predicate;
 
-import static junit.framework.Assert.assertEquals;
+import static org.junit.Assert.assertEquals;
 import static org.atatec.trugger.reflection.Reflection.reflect;
 
 /**