+++ src/test/java/org/atatec/trugger/test/reflection/FieldsSelectorTest.java
@@ -21,8 +21,8 @@ import org.junit.Test;
 
 import javax.annotation.Resource;
 
-import static junit.framework.Assert.assertFalse;
-import static junit.framework.Assert.assertTrue;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 import static org.atatec.trugger.reflection.Reflection.reflect;
 import static org.junit.Assert.assertEquals;