+++ src/test/java/org/atatec/trugger/test/reflection/MethodsSelectorTest.java
@@ -19,7 +19,7 @@ package org.atatec.trugger.test.reflection;
 import org.atatec.trugger.test.Flag;
 import org.junit.Test;
 
-import static junit.framework.Assert.assertEquals;
+import static org.junit.Assert.assertEquals;
 import static org.atatec.trugger.reflection.Reflection.reflect;
 
 /**