+++ src/test/java/org/atatec/trugger/test/validation/MinValidatorTest.java
@@ -25,6 +25,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import static org.mockito.Mockito.when;
+
 /**
  * @author Marcelo Guimarães
  */
@@ -32,7 +34,7 @@ public class MinValidatorTest extends BaseValidatorTest<Min> {
 
   @Test
   public void testPositiveValue() {
-    map(10.0).to(constraint.value());
+    when(constraint.value()).thenReturn(10.0);
 
     assertValid(10);
     assertInvalid(9);
@@ -41,7 +43,7 @@ public class MinValidatorTest extends BaseValidatorTest<Min> {
 
   @Test
   public void testNegativeValue() {
-    map(-10.0).to(constraint.value());
+    when(constraint.value()).thenReturn(-10.0);
 
     assertValid(-10);
     assertValid(9);
@@ -51,7 +53,7 @@ public class MinValidatorTest extends BaseValidatorTest<Min> {
 
   @Test
   public void testTypes() {
-    map(2.0).to(constraint.value());
+    when(constraint.value()).thenReturn(2.0);
 
     Object[] validArray = new Object[2];
     Object[] invalidArray = new Object[1];
@@ -90,8 +92,8 @@ public class MinValidatorTest extends BaseValidatorTest<Min> {
 
   @Test
   public void testDelta() {
-    map(2.0).to(constraint.value());
-    map(1.0).to(constraint.delta());
+    when(constraint.value()).thenReturn(2.0);
+    when(constraint.delta()).thenReturn(1.0);
 
     assertValid(new Object[1]);
     assertInvalid(new Object[0]);
@@ -99,8 +101,8 @@ public class MinValidatorTest extends BaseValidatorTest<Min> {
 
   @Test
   public void testInclusive() {
-    map(2.0).to(constraint.value());
-    map(false).to(constraint.inclusive());
+    when(constraint.value()).thenReturn(2.0);
+    when(constraint.inclusive()).thenReturn(false);
 
     assertValid(new Object[3]);
     assertInvalid(new Object[2]);